requirements.txt
flask
flask-socketio
python-socketio
opencv-python
mediapipe
python-dotenv 